requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
celery
redis
python-dotenv
groq
requests
gunicorn
